+++ SpaceInvaders/app/src/main/java/com/example/aleja/spaceinvaders/Laser.java
@@ -2,10 +2,15 @@ package com.example.aleja.spaceinvaders;
 
 import android.graphics.RectF;
 
+import java.util.Random;
+
 public class Laser {
     protected float x;
     protected float y;
     protected RectF rect;
+
+    private boolean randomStatus;
+    private boolean letal;
 
     // En qué dirección se está disparando
     public final int ARRIBA = 0;
@@ -47,6 +52,13 @@ public class Laser {
 
     }
 
+    public boolean getRandomStatus(){
+        return this.randomStatus;
+    }
+    public void setRandomStatus(Boolean b) {
+        this.randomStatus = b;
+    }
+
     public float getX(){
         return this.x;
     }
@@ -58,6 +70,7 @@ public class Laser {
         if (!isActive) {
             x = startX;
             y = startY;
+            letal = false;   // Solo sirve para los laser de invaders
             heading = direction;
             isActive = true;
             return true;
@@ -90,5 +103,24 @@ public class Laser {
         rect.top = y;
         rect.bottom = y + height;
 
+    }
+
+    public void hacerLetal(){
+        this.letal = true;
+    }
+    public boolean isLetal() {
+        return this.letal;
+    }
+
+    public void randomMove(int ejeX, int percentaje){
+        Random randomGen = new Random();
+        int randomMoveStatus = randomGen.nextInt(100);
+        if (randomMoveStatus < percentaje){
+            this.setRandomStatus(true);
+            this.setX(randomGen.nextInt(ejeX));
+
+        }else{
+            this.setRandomStatus(false);
+        }
     }
 }
